CA-825 Change directory structure to have gradle sub-project per CRL package (#20)
We want to publish and test CRL packages independently of each other to match the cloud api package division. This is the first step, reworking our directory structure to use top-level directories per CRL package and using Gradle multi-project builds.

Also deletes 'resourcemanager' package now that we have other examples and we don't plan on using that cloud library. We'll re-add it later with the api client library dependency.